requirements.txt
astroid==2.4.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.1
pandas==1.1.0
psycopg2-binary==2.8.5
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
toml==0.10.1
wrapt==1.12.1
